requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
bleach==5.0.0
bump2version==1.0.1
bumpversion==0.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
commonmark==0.9.1
distlib==0.3.4
dj-database-url==0.5.0
Django==3.2.9
django-bona-blog==1.1.4
django-ckeditor==6.0.0
django-cleanup==2.1.0
django-crispy-forms==1.14.0
django-filter==21.1
django-froala-editor==4.0.10
django-js-asset==1.2.2
django-taggit==2.1.0
django-tinymce==3.4.0
django-tinymce4-lite==1.8.0
django-utils-six==2.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
docutils==0.18.1
filelock==3.6.0
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
joblib==1.1.0
jsmin==3.0.1
keyring==23.5.0
numpy==1.22.3
pandas==1.4.2
Pillow==8.4.0
pipenv==2022.3.28
pkginfo==1.8.2
platformdirs==2.5.1
promise==2.3
psycopg2==2.9.2
Pygments==2.11.2
python-dateutil==2.8.2
pytz==2021.3
pywin32-ctypes==0.2.0
readme-renderer==34.0
requests==2.27.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.2.0
Rx==1.6.1
scikit-learn==1.0.2
scipy==1.8.0
singledispatch==3.6.2
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
text-unidecode==1.3
threadpoolctl==3.1.0
twine==4.0.0
tzdata==2022.1
updates==0.1.7.1
urllib3==1.26.9
virtualenv==20.14.0
virtualenv-clone==0.5.7
webencodings==0.5.1
whitenoise==5.3.0
zipp==3.8.0
